requirements.txt
flask
openai
python-dotenv
reportlab
gunicorn
flask-limiter
